refactor: move settings into instance page
