+++ apps/app-frontend/src/components/ui/install_flow/AddServerToInstanceModal.vue
@@ -12,7 +12,7 @@ import { convertFileSrc } from '@tauri-apps/api/core'
 import { computed, ref } from 'vue'
 
 import ModalWrapper from '@/components/ui/modal/ModalWrapper.vue'
-import { instanceKeys } from '@/composables/instances/instance-query-options'
+import { instanceKeys } from '@/pages/instance/query-options'
 import { trackEvent } from '@/helpers/analytics'
 import { list } from '@/helpers/instance'
 import { add_server_to_instance, get_instance_worlds } from '@/helpers/worlds.ts'
