+++ apps/app-frontend/src/pages/instance/components/admonitions/shared-instance-stale.vue
