+++ apps/app-frontend/src/pages/instance/components/admonitions/shared-instance-update-available.vue
