+++ apps/app-frontend/src/pages/instance/components/page-header/instance-page-header-server-metadata.vue
