+++ apps/app-frontend/src/pages/instance/components/settings-modal/hooks-settings.vue
@@ -10,9 +10,9 @@ import { computed, ref, watch } from 'vue'
 
 import { edit } from '@/helpers/instance'
 import { get } from '@/helpers/settings.ts'
-import { injectInstanceSettings } from '@/providers/instance-settings'
+import { injectInstanceSettings } from './instance-settings-context'
 
-import type { AppSettings, Hooks } from '../../../helpers/types'
+import type { AppSettings, Hooks } from '../../../../helpers/types'
 
 const { handleError } = injectNotificationManager()
 const { formatMessage } = useVIntl()
