+++ apps/app-frontend/src/pages/instance/components/settings-modal/shared-instance-installation-settings-controls.vue
