+++ apps/app-frontend/src/pages/instance/components/settings-modal/sharing-settings.vue
@@ -94,9 +94,6 @@ import {
 import { useMutation, useQuery, useQueryClient } from '@tanstack/vue-query'
 import { computed, ref } from 'vue'
 
-import ConfirmRevokeSharedInstanceInviteModal from '@/components/ui/shared-instances/ConfirmRevokeSharedInstanceInviteModal.vue'
-import SharedInstanceInstallationSettingsControls from '@/components/ui/shared-instances/SharedInstanceInstallationSettingsControls.vue'
-import { instanceKeys } from '@/composables/instances/instance-query-options'
 import { config } from '@/config'
 import {
 	get_shared_instance_invites,
@@ -105,7 +102,11 @@ import {
 	unpublish_shared_instance,
 } from '@/helpers/instance'
 import { useSharedInstanceErrors } from '@/helpers/shared-instance-errors'
-import { injectInstanceSettings } from '@/providers/instance-settings'
+
+import { instanceKeys } from '../../query-options.ts'
+import ConfirmRevokeSharedInstanceInviteModal from './confirm-revoke-shared-instance-invite-modal.vue'
+import { injectInstanceSettings } from './instance-settings-context.ts'
+import SharedInstanceInstallationSettingsControls from './shared-instance-installation-settings-controls.vue'
 
 const { instance, offline, onUnlinked } = injectInstanceSettings()
 const { notifySharedInstanceError } = useSharedInstanceErrors()
