+++ apps/app-frontend/src/pages/instance/share/use-shared-instance-members.ts
@@ -2,7 +2,6 @@ import type { InvitePlayersUser } from '@modrinth/ui'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/vue-query'
 import { computed, type Ref, ref } from 'vue'
 
-import { instanceKeys } from '@/composables/instances/instance-query-options'
 import { get_user_many } from '@/helpers/cache.js'
 import {
 	get_shared_instance_users,
@@ -13,6 +12,7 @@ import {
 } from '@/helpers/instance'
 import type { GameInstance } from '@/helpers/types'
 
+import { instanceKeys } from '../query-options'
 import {
 	normalizeInviteKey,
 	SHARED_INSTANCE_USER_LIMIT,
